ProfitClient: ramp policy falls through to defaults when env unset; relocate predicate
Address pxrl feedback on #3468:

- Move isUnmeteredFastRebalance out of a one-off FastRebalanceUtils.ts into the
  existing FillUtils.ts and drop the standalone file.
- In resolveGasMultiplier and minRelayerFeePct, only short-circuit to the
  RELAYER_RAMP_* value when the env var is actually defined. If the operator
  configures destinations/origins but leaves the override unset, fall through
  to the standard per-route/token/chain lookup and global default rather than
  silently using 0.
- Update the README and test to match.
