+++ src/clients/ProfitClient.ts
@@ -180,8 +180,12 @@ export class ProfitClient {
     const srcSymbol = this.getTokenSymbol(inputToken, originChainId);
     const dstSymbol = this.getTokenSymbol(outputToken, destinationChainId);
 
-    if (this.matchesRampPolicy(srcSymbol, dstSymbol, originChainId, destinationChainId, inputToken)) {
-      const rampMultiplier = toBNWei(process.env.RELAYER_RAMP_GAS_MULTIPLIER ?? "0");
+    const rampGasMultiplier = process.env.RELAYER_RAMP_GAS_MULTIPLIER;
+    if (
+      isDefined(rampGasMultiplier) &&
+      this.matchesRampPolicy(srcSymbol, dstSymbol, originChainId, destinationChainId, inputToken)
+    ) {
+      const rampMultiplier = toBNWei(rampGasMultiplier);
       assert(
         rampMultiplier.gte(bnZero) && rampMultiplier.lte(toBNWei(4)),
         `RELAYER_RAMP_GAS_MULTIPLIER out of range (${rampMultiplier})`
@@ -429,8 +433,12 @@ export class ProfitClient {
     const srcSymbol = this.getTokenSymbol(inputToken, originChainId);
     const dstSymbol = this.getTokenSymbol(outputToken, destinationChainId);
 
-    if (this.matchesRampPolicy(srcSymbol, dstSymbol, originChainId, destinationChainId, inputToken)) {
-      return toBNWei(process.env.RELAYER_RAMP_MIN_FEE_PCT ?? "0");
+    const rampMinFeePct = process.env.RELAYER_RAMP_MIN_FEE_PCT;
+    if (
+      isDefined(rampMinFeePct) &&
+      this.matchesRampPolicy(srcSymbol, dstSymbol, originChainId, destinationChainId, inputToken)
+    ) {
+      return toBNWei(rampMinFeePct);
     }
 
     const effectiveSourceSymbol = this._getRemappedTokenSymbol(srcSymbol) ?? srcSymbol;
@@ -573,8 +581,9 @@ export class ProfitClient {
    * Per-(token-pair) override: a deposit matches when its destination is in
    * `RELAYER_RAMP_DESTINATIONS_<srcSymbol>_<dstSymbol>`, optionally its origin is in
    * `RELAYER_RAMP_ORIGINS_<srcSymbol>_<dstSymbol>`, and the origin chain supports unmetered fast
-   * rebalance for the input token. Matching deposits use `RELAYER_RAMP_MIN_FEE_PCT` and
-   * `RELAYER_RAMP_GAS_MULTIPLIER` instead of the per-route lookups.
+   * rebalance for the input token. Callers additionally gate on the relevant override env var being
+   * defined (`RELAYER_RAMP_MIN_FEE_PCT` / `RELAYER_RAMP_GAS_MULTIPLIER`); if it isn't, they fall
+   * through to the standard per-route/token/chain lookup and global defaults.
    */
   protected matchesRampPolicy(
     srcSymbol: string,
