+++ src/utils/FillUtils.ts
@@ -1,6 +1,8 @@
+import { utils as sdkUtils } from "@across-protocol/sdk";
+import { CHAIN_IDs, TOKEN_SYMBOLS_MAP } from "@across-protocol/constants";
 import { HubPoolClient, SpokePoolClient } from "../clients";
 import { FillStatus, FillWithBlock, SpokePoolClientsByChain, DepositWithBlock, RelayData } from "../interfaces";
-import { EMPTY_MESSAGE } from "../utils";
+import { Address, compareAddressesSimple, EMPTY_MESSAGE } from "../utils";
 
 export type RelayerUnfilledDeposit = {
   deposit: DepositWithBlock;
@@ -64,6 +66,25 @@ export function getUnfilledDeposits(
       const version = hubPoolClient.configStoreClient.getConfigStoreVersionForTimestamp(deposit.quoteTimestamp);
       return { deposit, version, invalidFills };
     });
+}
+
+/**
+ * Returns true if `(chainId, token)` can be drained off via an unmetered, low-latency bridge
+ * (CCTP for USDC, OFT for USDT, or the hub chain itself via canonical bridges).
+ *
+ * Pure over the static CCTP/OFT chain registries plus a single hub-chain id, so it's safe to
+ * call without taking a dependency on InventoryClient.
+ */
+export function isUnmeteredFastRebalance(chainId: number, token: Address, hubChainId: number): boolean {
+  const cctp =
+    sdkUtils.chainIsCCTPEnabled(chainId) &&
+    compareAddressesSimple(TOKEN_SYMBOLS_MAP.USDC.addresses[chainId], token.toNative());
+  // OFT withdrawals from HyperEVM take ~12h, so they aren't fast.
+  const oft =
+    sdkUtils.chainIsOFTEnabled(chainId) &&
+    compareAddressesSimple(TOKEN_SYMBOLS_MAP.USDT.addresses[chainId], token.toNative()) &&
+    chainId !== CHAIN_IDs.HYPEREVM;
+  return cctp || oft || chainId === hubChainId;
 }
 
 export function depositForcesOriginChainRepayment(
