+++ src/utils/index.ts
@@ -87,4 +87,3 @@ export * from "./TimeUtils";
 export * from "./DepositAddressUtils";
 export * from "./BridgeUtils";
 export * from "./RunningBalanceUtils";
-export * from "./FastRebalanceUtils";
